chore: move dummy components to UIShared directory
