+++ src/components/AdminSettings/TurnServers.vue
@@ -71,7 +71,7 @@ import { loadState } from '@nextcloud/initial-state'
 import NcButton from '@nextcloud/vue/dist/Components/NcButton.js'
 
 import TurnServer from '../../components/AdminSettings/TurnServer.vue'
-import TransitionWrapper from '../TransitionWrapper.vue'
+import TransitionWrapper from '../UIShared/TransitionWrapper.vue'
 
 export default {
 	name: 'TurnServers',
