+++ src/components/LeftSidebar/ConversationsList/ConversationsSearchListVirtual.vue
@@ -39,7 +39,7 @@
 import { RecycleScroller } from 'vue-virtual-scroller'
 
 import ConversationSearchResult from './ConversationSearchResult.vue'
-import LoadingPlaceholder from '../../LoadingPlaceholder.vue'
+import LoadingPlaceholder from '../../UIShared/LoadingPlaceholder.vue'
 
 import 'vue-virtual-scroller/dist/vue-virtual-scroller.css'
