+++ src/components/LeftSidebar/LeftSidebar.vue
@@ -316,11 +316,11 @@ import Conversation from './ConversationsList/Conversation.vue'
 import ConversationsListVirtual from './ConversationsList/ConversationsListVirtual.vue'
 import InvitationHandler from './InvitationHandler.vue'
 import OpenConversationsList from './OpenConversationsList/OpenConversationsList.vue'
-import SearchBox from './SearchBox/SearchBox.vue'
+import SearchBox from '../UIShared/SearchBox.vue'
 import ConversationIcon from '../ConversationIcon.vue'
-import Hint from '../Hint.vue'
+import Hint from '../UIShared/Hint.vue'
 import NewConversationDialog from '../NewConversationDialog/NewConversationDialog.vue'
-import TransitionWrapper from '../TransitionWrapper.vue'
+import TransitionWrapper from '../UIShared/TransitionWrapper.vue'
 
 import { useArrowNavigation } from '../../composables/useArrowNavigation.js'
 import { ATTENDEE, CONVERSATION } from '../../constants.js'
