+++ src/components/UIShared/SearchBox.vue
