+++ src/components/UIShared/VolumeIndicator.vue
