+++ src/views/MainView.vue
@@ -20,7 +20,7 @@ import ChatView from '../components/ChatView.vue'
 import LobbyScreen from '../components/LobbyScreen.vue'
 import PollViewer from '../components/PollViewer/PollViewer.vue'
 import TopBar from '../components/TopBar/TopBar.vue'
-import TransitionWrapper from '../components/TransitionWrapper.vue'
+import TransitionWrapper from '../components/UIShared/TransitionWrapper.vue'
 
 import { useIsInCall } from '../composables/useIsInCall.js'
